requirements.txt
python-telegram-bot==21.3
yt-dlp
psycopg2-binary
httpx
